+++ src/lime.js
@@ -2,14 +2,20 @@
 
 const { Client } = require('discord.js');
 const MessageSanitizer = require('./lib/message-sanitizer');
+const Stress = require('./lib/stress');
+const User = require('./lib/user');
 const botModules = require('./modules');
 
 module.exports = class Lime {
 
-	constructor({ discordToken } = {}) {
+	constructor(config) {
+
 		this.discordClient = new Client();
-		this.discordToken = discordToken;
+		this.stress = new Stress(config.stress);
+
+		this.config = config;
 		this.modules = {};
+		this.usersById = {};
 	}
 
 	async initializeModules() {
@@ -31,13 +37,30 @@ module.exports = class Lime {
 
 		this.discordClient.on('message', message => this.handleMessage(message));
 
-		this.discordClient.login(this.discordToken);
+		this.discordClient.login(this.config.discordToken);
 	}
 
 	async handleMessage(message) {
 
 		if(!message.mentions.has(this.discordClient.user))
 			return;
+
+		const user = this.getUserById(message);
+
+		if(!user.stress.pressure()) {
+			console.log(`Bot currently tired of ${user.name}: stress level ${user.stress.level}`);
+			return;
+		}
+
+		if(!this.stress.pressure()) {
+			console.log(`Bot currently stressed: stress level ${this.stress.level}`);
+			return;
+		}
+
+		console.log({
+			mainStress: this.stress.level,
+			userStress: user.stress.level
+		});
 
 		message.content = MessageSanitizer.clean(
 			[
@@ -48,10 +71,18 @@ module.exports = class Lime {
 			message.content
 		);
 
-		console.log('Replying to:', message.content);
+		console.log(`Replying to ${user.tag}:`, message.content);
 
-		const response = await this.modules.chatbot.chat(message);
+		const response = await this.modules.chatbot.chat(message, user);
 
 		return message.reply(response);
+	}
+
+	getUserById({ author, member }) {
+
+		if(!this.usersById[author.id])
+			this.usersById[author.id] = new User(author, member, this.config.user.stress);
+
+		return this.usersById[author.id];
 	}
 };