+++ src/modules/chatbot/lib/dynamic-formatters/index.js
@@ -1,7 +1,9 @@
 'use strict';
 
 const timeFormatter = require('./time');
+const userFormatter = require('./user');
 
 module.exports = {
-	time: timeFormatter
+	time: timeFormatter,
+	user: userFormatter
 };