+++ src/modules/chatbot/lib/user.js
@@ -1,8 +1,0 @@
-'use strict';
-
-module.exports = class User {
-
-	constructor(userName) {
-		this.userName = userName;
-	}
-};